+++ src/materials/__init__.py
@@ -1,0 +1,5 @@
+from .aluminium import aluminium as aluminium
+from .steel import steel as steel
+from .pet import pet as pet
+from .pp import pp as pp
+from .rubber import rubber as rubber
